add gates to the project

+++ app/Http/Controllers/RoleController.php
@@ -16,7 +16,7 @@ class RoleController extends Controller
      */
     public function index()
     {
-        //Gate::authorize('haveAccess', 'role.index');
+        Gate::authorize('haveAccess', 'role.index');
         $roles = Role::orderBy('id', 'Desc')->paginate(3);
         
         return view('role.index', compact('roles'));
@@ -29,6 +29,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('haveAccess', 'role.create');
         $permissions = Permission::get();
 
         return view('role.create', compact('permissions'));
@@ -42,6 +43,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('haveAccess', 'role.create');
         $request->validate([
             'name' => 'required|max:50|unique:roles,name',
             'slug' => 'required|max:50|unique:roles,slug',
@@ -66,6 +68,7 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        $this->authorize('haveAccess', 'role.show');
         $permission_role = [];
 
         foreach($role->permissions as $permission) {
@@ -85,6 +88,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('haveAccess', 'role.edit');
         $permission_role = [];
 
         foreach($role->permissions as $permission) {
@@ -105,6 +109,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        $this->authorize('haveAccess', 'role.edit');
         $request->validate([
             'name' => 'required|max:50|unique:roles,name,'.$role->id,
             'slug' => 'required|max:50|unique:roles,slug,'.$role->id,
@@ -129,6 +134,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        $this->authorize('haveAccess', 'role.destroy');
         $role->delete();
 
         return redirect()->route('role.index')->with('status_success', 'Role  Succesfully removed');
